Add files via upload
One idea to consider: when encountering incorrect light sources, like light leaking through walls, it’s possible that the developers originally intended for it to be that way. Another way to approach the fix could be to remove the unwanted light (as I did) but also add a replacement light source—like a torch sconce—to maintain the intended ambiance.

Regardless of the approach, this was a fun and interesting challenge to tackle.